% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi_bio_naive.R
\name{trouver_sequence_naive_bio}
\alias{trouver_sequence_naive_bio}
\title{Algorithme naïf pour séquence ADN}
\usage{
trouver_sequence_naive_bio(obs, states, pi, A, B)
}
\arguments{
\item{obs}{Séquence d'observations (codée en entiers correspondant aux bases A, T, C, G)}

\item{states}{États possibles (vecteur de chaînes, ex: c("C", "N"))}

\item{pi}{Vecteur de probabilité initiale}

\item{A}{Matrice de transition}

\item{B}{Matrice d’émission}
}
\value{
Meilleure séquence d’états (vecteur de caractères)
}
\description{
Applique la méthode exhaustive pour trouver la meilleure séquence d'états cachés
}
