% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viterbi.R
\name{viterbi_bio_optimise}
\alias{viterbi_bio_optimise}
\title{Algorithme de Viterbi}
\usage{
viterbi_bio_optimise(obs, pi, A, B)
}
\arguments{
\item{obs}{Séquence d'observations (entiers).}

\item{pi}{Vecteur de probabilités initiales.}

\item{A}{Matrice de transition.}

\item{B}{Matrice d'émission.}
}
\value{
Une liste contenant la probabilité maximale et le chemin optimal.
}
\description{
Algorithme de Viterbi
}
